% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFormals.R
\name{assignFormals}
\alias{assignFormals}
\title{Extract and assign the default values of arguments (formals) of a function}
\usage{
assignFormals(fun, envir = .GlobalEnv)
}
\arguments{
\item{fun}{Function from which to extract defaults}

\item{envir}{Environment in which to assign the default values of the arguments}
}
\value{
Nothing (\code{invisible()}). The function is only used for its side-effects.
}
\description{
...and assign them to objects with the same names in an environment (global is default).
This is useful for testing and developing a function by running its source code line by line.
}
\author{
Kasper Skytte Andersen
}
