% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colClasses-set.R
\name{colClasses<-}
\alias{colClasses<-}
\title{Set (coerce) the column classes of a dataframe}
\usage{
colClasses(df) <- value
}
\arguments{
\item{df}{Dataframe like object}

\item{value}{Character vector containing elements of: \code{"c"} (character), \code{"n"} (numeric), \code{"f"} (factor), \code{"l"} (logical)" to set the column classes of \code{df}.}
}
\value{
A dataframe with coerced columns
}
\description{
Set (coerce) the column classes of a dataframe
}
\examples{
df <- iris[1:10, ]
df
colClasses(df)
colClasses(df) <- "nncfc"
colClasses(df)
}
