% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractImports.R
\name{extractImports}
\alias{extractImports}
\title{Extract all unique packages listed in a namespace file}
\usage{
extractImports(namespace = "NAMESPACE")
}
\arguments{
\item{namespace}{Path to \code{NAMESPACE} file. (\emph{default:} \code{"NAMESPACE"})}
}
\value{
A character vector with the packages found
}
\description{
Extract all unique packages listed in a namespace file
}
