% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_colClasses-set.R
\name{ksa_colClasses<-}
\alias{ksa_colClasses<-}
\title{Set (coerce) the column classes of a dataframe}
\usage{
ksa_colClasses(df) <- value
}
\arguments{
\item{df}{Dataframe like object}

\item{value}{Character vector containing elements of: \code{"c"} (character), \code{"n"} (numeric), \code{"f"} (factor), \code{"l"} (logical)" to set the column classes of \code{df}.}
}
\value{
A dataframe with coerced columns
}
\examples{
df <- iris[1:10,]
df
ksa_colClasses(df)
ksa_colClasses(df) <- "nncfc"
ksa_colClasses(df)
}
