% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_colClasses.R
\name{ksa_colClasses}
\alias{ksa_colClasses}
\title{Column classes of a data frame}
\usage{
ksa_colClasses(df)
}
\arguments{
\item{df}{A data frame like object}
}
\value{
A data frame containing the column classes of the provided data frame.
}
\description{
Returns a vector with the column classes of a data frame like object.
}
\note{
Cannot handle columns with more than one class.
}
\examples{
ksa_colClasses(iris)
}
