% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_funsInFile.R
\name{ksa_funsInFile}
\alias{ksa_funsInFile}
\title{List all R functions used in a file}
\usage{
ksa_funsInFile(filename, alphabetic = TRUE, include_base = FALSE)
}
\arguments{
\item{include_base}{}
}
\description{
List all R functions used in a file
}
