% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_gogoawesomepkg.R
\name{ksa_gogoawesomepkg}
\alias{ksa_gogoawesomepkg}
\title{Document, check, and install a package}
\usage{
ksa_gogoawesomepkg(install = TRUE, pkgdown = TRUE, verbose = TRUE)
}
\arguments{
\item{install}{Install the package (\code{TRUE}) with \code{\link[devtools]{install}} or just load it in memory (\code{FALSE}) with \code{\link[devtools]{load_all}}. (Default: \code{TRUE})}

\item{pkgdown}{Build package documentation website or not. (Default: \code{TRUE})}

\item{verbose}{Print status messages during the process or not. (Default: \code{TRUE})}
}
\value{
Invisibly returns a list with the output of all steps run.
}
\description{
Everything I do before commiting changes to GitHub. Generates package man pages, checks, installs, and builds package documentation website etc.
}
