% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_pkgDocs.R
\name{ksa_pkgDocs}
\alias{ksa_pkgDocs}
\title{Generate package documentation and webpage}
\usage{
ksa_pkgDocs()
}
\description{
Just runs \code{roxygen2::roxygenize(); pkgdown::clean_site(); pkgdown::build_site()}. I ran this so many times now that a small function like this is actually useful.
}
