% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksatools.R
\docType{package}
\name{ksatools}
\alias{ksatools}
\title{\code{ksatools} package}
\description{
Tools to make my R and package development life easier.
}
\details{
Please use the online documentation webpage at \href{https://kasperskytte.github.io/ksatools/}{https://kasperskytte.github.io/ksatools/}
}
