% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressAll.R
\name{suppressAll}
\alias{suppressAll}
\title{Suppress all types of diagnostic messages, error messages, and/or outputs}
\usage{
suppressAll(expr)
}
\arguments{
\item{expr}{Expression to be evaluated.}
}
\description{
Simply wraps the given expression in \code{\link{tryCatch}}, \code{\link{invisible}}, \code{\link{capture.output}}, \code{\link{suppressMessages}}, \code{\link{suppressWarnings}}, and \code{\link{suppressPackageStartupMessages}} to suppress any possible console output.
}
\examples{
suppressAll({
  stop("Some error")
  print("Some print")
  message("Some message")
  warning("Some warning")
  some_variable <- "test"
  return(some_variable)
})
}
