% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_gogoawesomepkg.R
\name{theactualgogoawesomepkgfunction}
\alias{theactualgogoawesomepkgfunction}
\title{gogo}
\usage{
theactualgogoawesomepkgfunction(style = TRUE, check = TRUE,
  install = TRUE, pkgdown = TRUE, verbose = TRUE,
  num_threads = parallel::detectCores() - 2L)
}
\arguments{
\item{style}{Passed on from \code{\link{ksa_gogoawesomepkg}}}

\item{check}{Passed on from \code{\link{ksa_gogoawesomepkg}}}

\item{install}{Passed on from \code{\link{ksa_gogoawesomepkg}}}

\item{pkgdown}{Passed on from \code{\link{ksa_gogoawesomepkg}}}

\item{verbose}{Passed on from \code{\link{ksa_gogoawesomepkg}}}

\item{num_threads}{Passed on from \code{\link{ksa_gogoawesomepkg}}}
}
\value{
See \code{\link{ksa_gogoawesomepkg}}
}
\description{
See \code{\link{ksa_gogoawesomepkg}}
}
