% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_colClasses.R
\name{ksa_colClasses}
\alias{ksa_colClasses}
\title{Get the column classes of a dataframe}
\usage{
ksa_colClasses(df)
}
\arguments{
\item{df}{Dataframe like object}
}
\value{
A dataframe containing the column classes of the provided dataframe
}
\description{
Returns a vector with the column classes of a dataframe like object, or coerces the column classes as defined by user.
}
\note{
Cannot handle columns with more than one class.
}
\examples{
df <- iris[1:10,]
df
ksa_colClasses(df)
}
