% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksa_colClasses.R
\name{ksa_colClasses}
\alias{ksa_colClasses}
\title{Column classes}
\usage{
ksa_colClasses(df)
}
\arguments{
\item{df}{A data frame like object}
}
\value{
A character vector with the column classes of the provided data frame.
}
\description{
Returns a vector with the column classes of a data frame like object.
}
\examples{
ksa_colClasses(iris)
}
