% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogoawesomepkg.R
\name{theactualgogoawesomepkgfunction}
\alias{theactualgogoawesomepkgfunction}
\title{gogo}
\usage{
theactualgogoawesomepkgfunction(
  style = TRUE,
  check = TRUE,
  test = FALSE,
  install = TRUE,
  pkgdown = TRUE,
  verbose = TRUE,
  num_threads = parallel::detectCores() - 2L
)
}
\arguments{
\item{style}{Passed on from \code{\link{gogoawesomepkg}}}

\item{check}{Passed on from \code{\link{gogoawesomepkg}}}

\item{test}{Passed on from \code{\link{gogoawesomepkg}}}

\item{install}{Passed on from \code{\link{gogoawesomepkg}}}

\item{pkgdown}{Passed on from \code{\link{gogoawesomepkg}}}

\item{verbose}{Passed on from \code{\link{gogoawesomepkg}}}

\item{num_threads}{Passed on from \code{\link{gogoawesomepkg}}}
}
\value{
See \code{\link{gogoawesomepkg}}
}
\description{
See \code{\link{gogoawesomepkg}}
}
\keyword{internal}
